Bug 1035892 - Handle 64-bit addresses for EXC_BAD_ACCESS exceptions on Mac r=froydnj
Differential Revision: https://phabricator.services.mozilla.com/D24356

+++ toolkit/crashreporter/breakpad-client/mac/handler/minidump_generator.h
@@ -97,7 +97,7 @@ class MinidumpGenerator {
   bool Write(const char *path);
 
   // Specify some exception information, if applicable
-  void SetExceptionInformation(int type, int code, int subcode,
+  void SetExceptionInformation(int type, int code, int64_t subcode,
                                mach_port_t thread_name) {
     exception_type_ = type;
     exception_code_ = code;
@@ -200,7 +200,7 @@ class MinidumpGenerator {
   // Exception information
   int exception_type_;
   int exception_code_;
-  int exception_subcode_;
+  int64_t exception_subcode_;
   mach_port_t exception_thread_;
   mach_port_t crashing_task_;
   mach_port_t handler_thread_;
